+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsefunction.py
@@ -507,9 +507,9 @@ class EvaluatorResponseFunction(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-07-08T21:19:26.866Z"
+    created: Optional[str] = "2026-07-09T07:44:19.589Z"
 
-    updated: Optional[str] = "2026-07-08T21:19:26.866Z"
+    updated: Optional[str] = "2026-07-09T07:44:19.589Z"
 
     updated_by_id: OptionalNullable[str] = UNSET