+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsepython.py
@@ -40,9 +40,9 @@ class EvaluatorResponsePython(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-07-08T21:19:26.866Z"
+    created: Optional[str] = "2026-07-09T07:44:19.589Z"
 
-    updated: Optional[str] = "2026-07-08T21:19:26.866Z"
+    updated: Optional[str] = "2026-07-09T07:44:19.589Z"
 
     updated_by_id: OptionalNullable[str] = UNSET