+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KX1SKS11V81TWG0EWQW7VJQE"
+        "01KX2XC0Q8D0HJ8KEHT5NC367X"
     )
     r"""The unique identifier of the data source"""