+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KX1SKS13K30R5AQN1J55KHFK"
+        "01KX2XC0QAGZP1Z3GDFWYR4K6M"
     )
     r"""The unique identifier of the data source"""