+++ packages/orq-rc/RELEASES.md
@@ -9369,3 +9369,13 @@ Based on:
 - [python v4.12.0-rc.9] packages/orq-rc
 ### Releases
 - [PyPI v4.12.0-rc.9] https://pypi.org/project/orq-ai-sdk/4.12.0-rc.9 - packages/orq-rc
+
+## 2026-07-09 07:50:29
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.789.0 (2.916.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.12.0-rc.10] packages/orq-rc
+### Releases
+- [PyPI v4.12.0-rc.10] https://pypi.org/project/orq-ai-sdk/4.12.0-rc.10 - packages/orq-rc